The ring and segment delete of active radar has been fixed (#682)
The ring and segment delete of active radar has been fixed

+++ src/main/java/com/a5lab/axion/domain/ring/RingServiceImpl.java
@@ -3,6 +3,7 @@ package com.a5lab.axion.domain.ring;
 import jakarta.persistence.criteria.Predicate;
 import java.util.ArrayList;
 import java.util.Collection;
+import java.util.LinkedList;
 import java.util.List;
 import java.util.Optional;
 import java.util.stream.Collectors;
@@ -14,6 +15,9 @@ import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Sort;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+
+import com.a5lab.axion.domain.ModelError;
+import com.a5lab.axion.domain.ValidationException;
 
 @RequiredArgsConstructor
 @Service
@@ -67,7 +71,12 @@ public class RingServiceImpl implements RingService {
   public void deleteById(Long id) {
     Optional<Ring> ringOptional = ringRepository.findById(id);
     if (ringOptional.isPresent()) {
-      new RadarActiveApprover(messageSource, ringOptional.get()).approve();
+      List<ModelError> modelErrorList = new LinkedList<>();
+      modelErrorList.addAll(new RadarActiveApprover(messageSource, ringOptional.get()).approve());
+      if (!modelErrorList.isEmpty()) {
+        String errorMessage = ValidationException.buildErrorMessage(modelErrorList);
+        throw new ValidationException(errorMessage, modelErrorList);
+      }
       ringRepository.deleteById(id);
     }
   }
